Add files via upload
Schematic 6 Stage Led Chaser Using Transistor

 

Thanks to the filling and discharging times of the transistors and capacitors used, the LEDs light up gradually and in a transitional manner, revealing a pleasant appearance. You only need to run it with 9V.

 

BC547 

Transistors, which are semiconductor circuit elements, are used as switching elements. Transistors with 3 or more legs are also preferred to increase voltage and current.



BC557?

BC557 is a BJT PNP type Transistor. The BC557 Transistor has a collector current of 100mA and a Collector-Emitter voltage of 45V. In order to see the functions and functions of the BC557 Transistor in the BC Transistor Series in more detail, be sure to examine the datasheet of this transistor.



Components

-100K Resistor

-1K Resistor

-BC547

-BC557

-22uF Capacitor